@@ -2,6 +2,7 @@
   "name": "apicarto",
   "version": "v2.7.4",
   "main": "index.js",
+  "type": "module",
   "scripts": {
     "test": "npm run lint && npm run test-unit",
     "test-unit": "NODE_ENV=test nyc --reporter=html --reporter=text --reporter=cobertura mocha",
@@ -25,31 +26,31 @@
     "@terraformer/wkt": "^2.1.2",
     "@turf/meta": "6.5.0",
     "@turf/turf": "^6.5.0",
-    "axios": "^1.6.5",
+    "axios": "^1.6.8",
     "body-parser": "^1.20.0",
     "bunyan": "^1.8.15",
     "codes-postaux": "^4.0.0",
     "cors": "^2.8.5",
     "debug": "^4.3.4",
-    "ejs": "^3.1.8",
-    "express": "^4.18.1",
+    "ejs": "^3.1.10",
+    "express": "^4.19.2",
     "express-validator": "^7.0.1",
     "geoportal-wfs-client": "https://github.com/IGNF/geoportal-wfs-client#v1.0.0",
     "handlebars": "^4.7.7",
-    "http-proxy-agent": "^5.0.0",
-    "https-proxy-agent": "^5.0.1",
+    "http-proxy-agent": "^7.0.2",
+    "https-proxy-agent": "^7.0.4",
     "lodash": "^4.17.21",
-    "pg": "^8.8.0",
+    "pg": "^8.11.5",
     "pg-format": "^1.0.4",
     "shelljs": "^0.8.5",
-    "swagger-ui-dist": "^5.11.0",
+    "swagger-ui-dist": "^5.16.2",
     "uuid": "^9.0.1"
   },
   "devDependencies": {
     "coveralls": "^3.1.1",
-    "eslint": "^8.23.0",
+    "eslint": "^9.1.0",
     "expect.js": "^0.3.1",
-    "mocha": "^10.0.0",
+    "mocha": "^10.4.0",
     "mocha-lcov-reporter": "^1.3.0",
     "nyc": "^15.1.0",
     "supertest": "^6.2.4"
