@@ -1,15 +1,13 @@
-var Router = require('express').Router;
+import { Router } from 'express';
+import cors from 'cors';
+import { check, matchedData } from 'express-validator';
+import validateParams from '../../middlewares/validateParams.js';
+import isGeometry from '../../checker/isGeometry.js';
+import gppWfsClient from '../../middlewares/gppWfsClient.js';
+import _ from 'lodash';
+
+
 var router = new Router();
-var cors = require('cors');
-const { check } = require('express-validator');
-const { matchedData } = require('express-validator');
-
-const validateParams = require('../../middlewares/validateParams');
-const {isGeometry} = require('../../checker');
-
-const gppWfsClient = require('../../middlewares/gppWfsClient');
-
-const _ = require('lodash');
 const lastYearRPG = 2022;
 const firstYearRPG = 2010;
 
@@ -116,4 +114,4 @@ router.post('/v2', cors(corsOptionsGlobal),rpgValidators, createRpgProxy('V2'));
 
 
 
-module.exports=router;
+export {router};
