@@ -1,6 +1,5 @@
-
-const  WKT = require('@terraformer/wkt');
-var flip = require('@turf/flip');
+import { geojsonToWKT } from '@terraformer/wkt';
+import flip from '@turf/flip';
 
 /*
  * WARNING: Despite the use of WGS84, you need to do a flip on the coordinates
@@ -54,7 +53,7 @@ function buildErCqlFilter(params) {
             if ( typeof geom !== 'object' ){
                 geom = JSON.parse(geom) ;
             }
-            var wkt = WKT.geojsonToWKT(flip(geom));
+            var wkt = geojsonToWKT(flip(geom));
             parts.push('INTERSECTS(the_geom,'+wkt+')');
         
         // Field_date sous la forme date_maj_deb;date_maj_fin, nous devons donc les separer pour la requete
@@ -95,4 +94,5 @@ function buildErCqlFilter(params) {
     }
     return parts.join(' and ') ;
 }
-module.exports = buildErCqlFilter;
+
+export default buildErCqlFilter;
