@@ -1,6 +1,5 @@
-
-const WKT = require('@terraformer/wkt');
-var flip = require('@turf/flip');
+import { geojsonToWKT } from '@terraformer/wkt';
+import flip from '@turf/flip';
 
 /*
  * WARNING: Despite the use of WGS84, you need to do a flip on the coordinates
@@ -28,7 +27,7 @@ function bboxToFilter(bbox){
     return 'BBOX(the_geom,'+xmin+','+ymin+','+xmax+','+ymax+')' ;
 }
 
-module.exports =  function(params){
+var cqlFilter =  function(params){
     var parts = [] ;
     for ( var name in params ){
         // ignore _limit, _start, etc.
@@ -43,7 +42,7 @@ module.exports =  function(params){
             if ( typeof geom !== 'object' ){
                 geom = JSON.parse(geom) ;
             }
-            var wkt = WKT.geojsonToWKT(flip(geom));
+            var wkt = geojsonToWKT(flip(geom));
             parts.push('INTERSECTS(geom,'+wkt+')');
         }else{
             parts.push(name+'=\''+ params[name]+'\'');
@@ -54,3 +53,5 @@ module.exports =  function(params){
     }
     return parts.join(' and ') ;
 };
+
+export default cqlFilter;
