@@ -1,8 +1,8 @@
-const GeoportalWfsClient = require('geoportal-wfs-client');
+import GeoportalWfsClient from 'geoportal-wfs-client';
 /*
  * middleware pour la création du client geoportail
  */
-module.exports = function(req, res, next) {
+var gpuWfsClient = function (req, res, next) {
     var referer = 'http://localhost';
 
     /* forward du referer du client */
@@ -20,3 +20,5 @@ module.exports = function(req, res, next) {
     });
     next();
 };
+
+export default gpuWfsClient;
