@@ -1,6 +1,8 @@
-module.exports = {
+var aocConfig = {
     version: '8 mars 2023',
     modules: ["Appellations viticoles"],
     nom_url :"FranceAgriMer",
     url: 'https://www.franceagrimer.fr/filieres-vin-et-cidre/vin/professionnels/teleprocedures'
 };
+
+export {aocConfig};
