@@ -1,6 +1,8 @@
-module.exports = {
+var banConfig = {
     version: 'v4.0.0',
     modules: ["Codes Postaux"],
     nom_url :"BAN",
     url: 'https://github.com/baseadressenationale/codes-postaux'
 };
+
+export {banConfig};
