@@ -1,6 +1,8 @@
-module.exports = {
+var geoportailConfig = {
     version: 'Flux WFS',
     modules: ["Cadastre", "RPG", "Nature", "WFS-Geoportail"],
     nom_url :"Geoservices",
     url: 'https://geoservices.ign.fr/services-web-experts'
 };
+
+export {geoportailConfig};
