@@ -1,6 +1,8 @@
-module.exports = {
+var gpuConfig = {
     version: 'Flux WFS',
     modules: ["GPU"],
     nom_url :"Géoportail de l'urbanisme",
     url: 'https://www.geoportail-urbanisme.gouv.fr'
 };
+
+export {gpuConfig};
